# swagger-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "신일 프로젝트 API",
    "description": "신일 프로젝트의 REST API 문서입니다. 사용자 인증, 제품 관리, 고객 관리, 공지사항 등의 기능을 제공합니다.",
    "version": "1.0.0",
    "contact": {
      "name": "신일 프로젝트 개발팀",
      "email": "dev@shinil.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "로컬 API 서버 (실제 테스트 가능)"
    },
    {
      "url": "https://your-domain.vercel.app/api",
      "description": "프로덕션 서버"
    },
    {
      "url": "http://localhost:3003/api",
      "description": "로컬 테스트 API 서버"
    },
    {
      "url": "http://localhost:8080",
      "description": "로컬 정적 파일 서버 (문서만)"
    }
  ],
...
  "paths": {
...
    "/api/auth": {
      "post": {
        "tags": ["인증"],
        "summary": "사용자 로그인",
        "description": "이메일과 비밀번호를 사용하여 사용자 인증을 수행합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "사용자 이메일 주소",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "사용자 비밀번호",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "인증 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Authentication successful"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "액세스 토큰"
                        },
                        "refreshToken": {
                          "type": "string",
                          "description": "리프레시 토큰"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "사용자 ID"
                            },
                            "email": {
                              "type": "string",
                              "description": "사용자 이메일"
                            },
                            "role": {
                              "type": "string",
                              "description": "사용자 역할",
                              "enum": ["admin", "user"]
                            },
                            "approvalStatus": {
                              "type": "string",
                              "description": "승인 상태",
                              "enum": ["pending", "approved", "rejected"]
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "lastSignIn": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        },
                        "expiresIn": {
                          "type": "string",
                          "example": "24h"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Email and password are required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid email or password"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "승인 대기 중",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Account is pending approval. Please contact administrator."
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "허용되지 않는 메서드",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Method not allowed. Use POST."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/verify-token": {
      "post": {
        "tags": ["인증"],
        "summary": "토큰 검증",
        "description": "액세스 토큰의 유효성을 검증합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "검증할 액세스 토큰",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 검증 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Token is valid"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "사용자 ID"
                            },
                            "email": {
                              "type": "string",
                              "description": "사용자 이메일"
                            },
                            "role": {
                              "type": "string",
                              "description": "사용자 역할"
                            },
                            "approvalStatus": {
                              "type": "string",
                              "description": "승인 상태"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "lastSignIn": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        },
                        "valid": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "토큰 누락",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token is required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "토큰 무효",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 토큰을 사용한 인증"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
} 